+++ R/map_layer_pointcloud.R
@@ -16,8 +16,8 @@ mapdeckPointcloudDependency <- function() {
 #' with a certain radius.
 #'
 #' @inheritParams add_scatterplot
-#' @param elevation column containing the elevation values
-#' @param radius value in pixels of each point
+#' @param elevation column containing the elevation values. Default 0
+#' @param radius value in pixels of each point. Default 10.
 #' @param light_settings list of light setting parameters. See \link{light_settings}
 #'
 #' @inheritSection add_polygon data
@@ -111,7 +111,7 @@ add_pointcloud <- function(
 	l[["polyline"]] <- force( polyline )
 	l[["elevation"]] <- force( elevation )
 	l[["fill_colour"]] <- force( fill_colour)
-	l[["fill_opacity"]] <- force( fill_opacity )
+	l[["fill_opacity"]] <- resolve_opacity( fill_opacity )
 	l[["tooltip"]] <- force( tooltip )
 	l[["id"]] <- force( id )
 	l[["na_colour"]] <- force( na_colour )