+++ R/map_layer_polygon.R
@@ -19,18 +19,18 @@ mapdeckPolygonDependency <- function() {
 #'
 #' @param polyline optional column of \code{data} containing the polylines, if using encoded polylines
 #' @param fill_colour column of \code{data} or hex colour for the fill colour.
-#' @param fill_opacity value between 0 and 255. Either a string specifying the
-#' column of \code{data} containing the fill opacity of each shape, or a single value
-#' to be applied to all the shapes
+#' @param fill_opacity Either a string specifying the column of \code{data}
+#' containing the opacity of each shape, or a single value in [0,255], or [0, 1),
+#' to be applied to all the shapes. Default 255.
 #' @param stroke_colour variable of \code{data} or hex colour for the stroke. If used,
 #' \code{elevation} is ignored.
-#' @param stroke_opacity value between 0 and 255. Either a string specifying the
-#' column of \code{data} containing the stroke opacity of each shape, or a single value
-#' to be applied to all the shapes
-#' @param stroke_width width of the stroke in meters. If used, \code{elevation} is ignored.
+#' @param stroke_opacity Either a string specifying the column of \code{data}
+#' containing the opacity of each shape, or a single value in [0,255], or [0, 1),
+#' to be applied to all the shapes. Default 255.
+#' @param stroke_width width of the stroke in meters. If used, \code{elevation} is ignored. Default 1.
 #' @param light_settings list of light setting parameters. See \link{light_settings}
 #' @param elevation the height the polygon extrudes from the map. Only available if neither
-#' \code{stroke_colour} or \code{stroke_width} are supplied.
+#' \code{stroke_colour} or \code{stroke_width} are supplied. Default 0
 #'
 #' @section data:
 #'
@@ -140,9 +140,9 @@ add_polygon <- function(
 	l[["polyline"]] <- force( polyline )
 	l[["stroke_colour"]] <- force( stroke_colour )
 	l[["stroke_width"]] <- force( stroke_width )
-	l[["stroke_opacity"]] <- force( stroke_opacity )
+	l[["stroke_opacity"]] <- resolve_opacity( stroke_opacity )
 	l[["fill_colour"]] <- force( fill_colour )
-	l[["fill_opacity"]] <- force( fill_opacity )
+	l[["fill_opacity"]] <- resolve_opacity( fill_opacity )
 	l[["elevation"]] <- force( elevation )
 	l[["tooltip"]] <- force( tooltip )
 	l[["id"]] <- force( id )