+++ inst/include/layers/text.hpp
@@ -6,10 +6,6 @@
 namespace mapdeck {
 namespace text {
 
-// Rcpp::StringVector text_columns = Rcpp::StringVector::create(
-// 	"polyline","tooltip","angle","alignment_baseline","anchor","size"
-// );
-
 std::unordered_map< std::string, std::string > text_colours = {
 	{ "fill_colour", "fill_opacity" }
 };