+++ src/arc.cpp
@@ -13,8 +13,10 @@ Rcpp::List arc_defaults(int n) {
 
 
 // [[Rcpp::export]]
-Rcpp::List rcpp_arc_geojson( Rcpp::DataFrame data,
-                     Rcpp::List params, Rcpp::StringVector geometry_columns ) {
+Rcpp::List rcpp_arc_geojson(
+		Rcpp::DataFrame data, Rcpp::List params,
+		Rcpp::StringVector geometry_columns, int digits
+	) {
 
 	int data_rows = data.nrows();
 
@@ -33,13 +35,16 @@ Rcpp::List rcpp_arc_geojson( Rcpp::DataFrame data,
 		data_rows,
 		parameter_exclusions,
 		geometry_columns,
-		true  // jsonify legend
+		true,  // jsonify legend
+		digits
 	);
 }
 
 // [[Rcpp::export]]
-Rcpp::List rcpp_arc_geojson_df( Rcpp::DataFrame data,
-                             Rcpp::List params, Rcpp::List geometry_columns ) {
+Rcpp::List rcpp_arc_geojson_df(
+		Rcpp::DataFrame data, Rcpp::List params,
+		Rcpp::List geometry_columns, int digits
+	) {
 
 	int data_rows = data.nrows();
 
@@ -58,7 +63,8 @@ Rcpp::List rcpp_arc_geojson_df( Rcpp::DataFrame data,
 		data_rows,
 		parameter_exclusions,
 		geometry_columns,
-		true  // jsonify legend
+		true,  // jsonify legend
+		digits
 	);
 }